Managing tokens anew moduke introduced and implemented
