+++ app/assets/content/config.js
@@ -1,0 +1,19 @@
+/**
+ * All styles related to application will be kept in this file
+ */
+
+export const styleConstant = {
+	themeColor: '#106bc8',
+	textColor: '#ffffff',
+	activeTabUnderLineColor: '#ffffff',
+	tabNameFontSize: 15,
+	appNameFontSize: 20,
+	addTaskButtonColor: '#106bc8',
+	addCategoryButtonColor: 'rgba(231,76,60,1)',
+	addTaskIconName: 'playlist-add',
+	addCategoryIconName: 'library-add'
+};
+
+export const appConfig = {
+
+};
