+++ app/assets/content/index.js
@@ -1,0 +1,2 @@
+export { styleConstant } from './config';
+export { globalTokens } from './text';
