+++ app/assets/content/text.js
@@ -1,0 +1,8 @@
+/**
+ * All application specific tokens will be kept in file,
+ * Mainly tokens are distributed here based on screen
+ */
+
+export const globalTokens = {
+	applicationTitle: 'Task Manager'
+}
