+++ app/assets/styles/sharedStyles.js
@@ -3,18 +3,7 @@
  * style constant variables
  */
 import { StyleSheet } from 'react-native';
-
-export const styleConstant = {
-	themeColor: '#106bc8',
-	textColor: '#ffffff',
-	activeTabUnderLineColor: '#ffffff',
-	tabNameFontSize: 15,
-	appNameFontSize: 20,
-	addTaskButtonColor: '#106bc8',
-	addCategoryButtonColor: 'rgba(231,76,60,1)',
-	addTaskIconName: 'playlist-add',
-	addCategoryIconName: 'library-add'
-};
+import { styleConstant } from './../content/config';
 
 export const commonStyles = StyleSheet.create({
 	actionButtonIcon: {
