+++ app/commonModules.js
@@ -2,7 +2,8 @@
  * Purpose of this file is to change commont things
  * like icons and stylesheet from single place for all modules
  */
-import { commonStyles, styleConstant } from './assets/styles/sharedStyles';
+import { commonStyles } from './assets/styles/sharedStyles';
+import { styleConstant, globalTokens } from './assets/content';
 import ActionButton from 'react-native-action-button';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 
@@ -10,5 +11,6 @@ export {
 	ActionButton,
 	commonStyles,
 	Icon,
-	styleConstant
+	styleConstant,
+	globalTokens
 }
